+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/services/SessionManagementService.java
@@ -22,7 +22,7 @@ import org.apache.juli.logging.Log;
 import org.apache.juli.logging.LogFactory;
 import org.wso2.carbon.context.CarbonContext;
 import org.wso2.carbon.core.AbstractAdmin;
-import org.wso2.carbon.identity.application.authentication.framework.cache.SessionContextCache;
+import org.wso2.carbon.identity.application.authentication.framework.ServerSessionManagementService;
 import org.wso2.carbon.identity.application.authentication.framework.context.SessionContext;
 import org.wso2.carbon.identity.application.authentication.framework.internal.FrameworkServiceDataHolder;
 import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
@@ -35,19 +35,16 @@ public class SessionManagementService extends AbstractAdmin {
 
     public boolean removeSession(String sessionId) {
 
-        if (StringUtils.isBlank(sessionId)) {
-            return false;
-        }
-        // Retrieve session information from cache in order to publish event
-        SessionContext sessionContext = FrameworkUtils.getSessionContextFromCache(sessionId);
-        terminateSession(sessionContext, sessionId);
-        return true;
+        ServerSessionManagementService serverSessionManagementService =
+                FrameworkServiceDataHolder.getInstance().getServerSessionManagementService();
+        return serverSessionManagementService.removeSession(sessionId);
     }
 
     /**
      *
      * Terminates the requested session, after validating whether the session belongs to the logged in user.
-     *
+     * @deprecated Use UserSessionManagementService to terminate the session of the current user
+     * {@link org.wso2.carbon.identity.application.authentication.framework.UserSessionManagementService#terminateSessionBySessionId(String, String)}
      * @param sessionId
      * @return
      */
@@ -73,29 +70,13 @@ public class SessionManagementService extends AbstractAdmin {
         if (username.equals(authenticatedUser.getUserName())
                 && userStoreDomain.equals(authenticatedUser.getUserStoreDomain())
                 && carbonContext.getTenantDomain().equals(authenticatedUser.getTenantDomain())) {
-            terminateSession(sessionContext, sessionId);
+            ServerSessionManagementService serverSessionManagementService =
+                    FrameworkServiceDataHolder.getInstance().getServerSessionManagementService();
+            return serverSessionManagementService.removeSession(sessionId);
         } else { // TODO : Handle federated scenario.
             log.warn(String.format("Trying to terminate a session which does not belong to logged in user (%s). " +
                     "This might be an attempt for a security breach", username));
             return false;
         }
-        return true;
-    }
-
-    // TODO : Session ID should be a property of the SessionContext.
-    private void terminateSession(SessionContext sessionContext, String sessionId) {
-
-        if (FrameworkServiceDataHolder.getInstance().getAuthnDataPublisherProxy() != null && FrameworkServiceDataHolder
-                .getInstance().getAuthnDataPublisherProxy().isEnabled(null) && sessionContext != null) {
-
-            Object authenticatedUserObj = sessionContext.getProperty(FrameworkConstants.AUTHENTICATED_USER);
-            AuthenticatedUser authenticatedUser = new AuthenticatedUser();
-            if (authenticatedUserObj != null) {
-                authenticatedUser = (AuthenticatedUser) authenticatedUserObj;
-            }
-            FrameworkUtils.publishSessionEvent(sessionId, null, null, sessionContext, authenticatedUser,
-                    FrameworkConstants.AnalyticsAttributes.SESSION_TERMINATE);
-        }
-        SessionContextCache.getInstance().clearCacheEntry(sessionId);
     }
 }
